tdf#------ sw: textboxes in group shapes - part 3
Fix WPG import. Group Shapes before were imported via
oox, but complex content were missing from them.
Problem:
The oox imports the shapes correctly but it is unable
to load complex textboxes, on the other hand Writerfilter
can load textboxes correctly but there is no shape handler.
In addittion, the main problem was that, first oox loads
the shape then writerfilter the content, but if there is
a grouped version of shapes this will not work, because,
the parent shape is not ready only the child one so after
the content loading the another member shape have to be
load. When the textbox content be processed, there is no
shape and textbox at all.
Sulotion:
This patch introduces new uno-property named textboxcontent.
This is neccessary, because while the parser reaches the
texbox content there the writerfilter creates a new textbox for
it, and collects these textboxes to a queue for later use.
After the shape import is ready, from the queue, the texboxes
will be attached to each shape.
Conclusion:
Many import problem can be solved by this. On the other hand
proper test needed, because there can be many new regressions.
Many test have to be fixed because of the new situation.
Known problems:
The formatting of the text is lost so it have to be fixed
according to the new situation.
groups embed into other group still unsopported.
Test is missing
WIP

Please note the export and ODF handling will be introduced
in futher commit(s).

+++ sw/source/core/unocore/unodraw.cxx
@@ -964,7 +964,13 @@ SwXShape::~SwXShape()
 
 uno::Any SwXShape::queryInterface( const uno::Type& aType )
 {
-    uno::Any aRet = SwTextBoxHelper::queryInterface(GetFrameFormat(), aType);
+    uno::Any aRet;
+    if (aType == cppu::UnoType<text::XText>::get() ||
+        aType == cppu::UnoType<text::XTextAppend>::get() ||
+        aType == cppu::UnoType<text::XTextRange>::get())
+        aRet = SwTextBoxHelper::queryInterface(GetFrameFormat(), aType,
+                                               SdrObject::getSdrObjectFromXShape(mxShape));
+
     if (aRet.hasValue())
         return aRet;
 
@@ -1153,13 +1159,19 @@ void SwXShape::setPropertyValue(const OUString& rPropertyName, const uno::Any& a
             }
             else if (pEntry->nWID == FN_TEXT_BOX)
             {
-                bool bValue(false);
-                aValue >>= bValue;
-                if (bValue)
-                    SwTextBoxHelper::create(pFormat, GetSvxShape()->GetSdrObject());
-                else
-                    SwTextBoxHelper::destroy(pFormat, GetSvxShape()->GetSdrObject());
-
+                if (pEntry->nMemberId == MID_IS_TEXTBOX)
+                {
+                    bool bValue(false);
+                    aValue >>= bValue;
+                    if (bValue)
+                        SwTextBoxHelper::create(pFormat, GetSvxShape()->GetSdrObject());
+                    else
+                        SwTextBoxHelper::destroy(pFormat, GetSvxShape()->GetSdrObject());
+                }
+                if (pEntry->nMemberId == MID_TEXTBOX_CONTENT)
+                {
+                    SwTextBoxHelper::setTextBox(pFormat, GetSvxShape()->GetSdrObject(), aValue.get<uno::Reference<text::XTextFrame>>());
+                }
             }
             else if (pEntry->nWID == RES_CHAIN)
             {
@@ -1507,11 +1519,23 @@ uno::Any SwXShape::getPropertyValue(const OUString& rPropertyName)
                 else if (pEntry->nWID == FN_TEXT_BOX)
                 {
                     auto pSvxShape = GetSvxShape();
-                    bool bValue = SwTextBoxHelper::isTextBox(
-                        pFormat, RES_DRAWFRMFMT,
-                        ((pSvxShape && pSvxShape->GetSdrObject()) ? pSvxShape->GetSdrObject()
-                                                                  : pFormat->FindRealSdrObject()));
-                    aRet <<= bValue;
+                    if (pEntry->nMemberId == MID_IS_TEXTBOX)
+                    {
+
+                        bool bValue = SwTextBoxHelper::isTextBox(
+                            pFormat, RES_DRAWFRMFMT,
+                            ((pSvxShape && pSvxShape->GetSdrObject()) ? pSvxShape->GetSdrObject()
+                                : pFormat->FindRealSdrObject()));
+                        aRet <<= bValue;
+                    }
+                    if (pEntry->nMemberId == MID_TEXTBOX_CONTENT)
+                    {
+                        aRet = SwTextBoxHelper::queryInterface(
+                            pFormat, cppu::UnoType<text::XText>::get(),
+                            (pSvxShape && pSvxShape->GetSdrObject()
+                                 ? pSvxShape->GetSdrObject()
+                                 : pFormat->FindRealSdrObject()));
+                    }
                 }
                 else if (pEntry->nWID == RES_CHAIN)
                 {
